Fix removing software breakpoints
When a breakpoint is removed, the ebase.bpts array is correctly updated,
but ebase.bpsave is not kept in sync. This results in wrong instructions
being executed when adding multiple breakpoints and then removing any
but the last. Fix this error by updating the ebase.bpsave array when
removing a breakpoint.